Removed Wijkanders.
Wijkanders doesn't work at the moment since the parser needs to be rewritten
to accomodate for the new HTML structure of the Wijkander's menu. Keeping the
module makes refactoring harder, so I'll remove it. You are free to
write a new parser and make a pull request to put it back.