Replace decodeutf8 with decodeUtf8With (#71)
This closes #68.

* Removed the function Util.safeGet, it's not used anymore.
* Removed unused imports in Model.Karen.
* Replaced decodeUtf8 with decodeUtf8With in Model.Wijkanders.

The reasoning behind the changes: `decodueUtf8` is not total. It throws
errors when it finds characters it doesn't understand and probably in
other cases too. I propose that we everywhere `decodeUtf8` is used
replace it with `decodeUtf8With ignore` which ignores all invalid
characters. This will worst case produce empty strings, but I think
that's the correct tradeoff for our use case. It's better for the user
experience to show a meal missing some characters than to not show it
at all.