Fixed the Wijkanders import.
Wijkanders replaced their <br> with <p></p> and suddenly we didn't parse
anything. This should be fixed in this commit which changes the
algorithm to process content until next date or a last phrase is
encountered. I guess this will work until Wijkanders changes something
again. We'll see, we'll see...