Fix Wijkander's friday bug
Before this fix the parser would take all meals from the next week too
but now we stop at the next date instead of the next consecutive date.
Hopefully this makes the parser mure resilient to for instance easter
and fridays.

There are two tests in this commit, to verify that the parser can
handle the input data that broke the parser before.

This commit closes #83 and it also closes #88.