Simplify logging by sending all logs to stdout
Before this change, the data updater logs were sent to a file set in the
configuration of the application, which more or less hid the logs there.
Meanwhile the access logs were sent to stdout. Now both kinds of logs
are sent to stdout which makes it much easier for the programs hosting
mat to read and parse the logs.

Close #63.