Untangle the `refresh` function
The purpose of this commit is to make the `update` function a bit easier
to read and use. We do this by:

- Introducing `createViewReference` which creates an `IORef` to a `View`.
- Moving the parts of the code that creates a view reference from
  `refresh` to `createViewReference`, thus giving it one less concern.
- The `refresh` function now takes a view reference and an update
  semaphor/lock/whatever it is as arguments and when the lock is
  unlocked it runs the `update` function. This is much easier to read
  and understand than having `refresh` returning a callback, which it
  did before.

This change does not throw any shade on any of the earlier developers of
this project, it only recognizes that I'm becoming less clever for each
passing day.