Flatten the main function
This should make it a bit easier to read while hopefully making it
easier to change in the future.

Use pattern matching to see if the user wants to the see the usage
instructions, this allows us to make use of one `case` expression
instead of a `case` expression and an `if` expression.

Do configuration reification pointfree style while using lenses.

Co-authored-by: Jacob "Jassob" Jonsson <jacob.t.jonsson@gmail.com>