chore: remove Stack
We can build without Stack and it feels nice to not be dependent on
another tool built on top of another tool etc.

Massage some versions of dependencies and some datetime formatting code
to make everything work with the versions of the dependencies installed.

Use `haskell-ci` to generate GitHub workflow that works with cabal.